setter form 2 made into component, unstable checkpoint
